refactor: SuccessModal은 파일 구조를 변경하여 보여지기 편해지도록 폴더 구조 수정
